Set Contract OrderLine Quantity as float instead of int to prevent division by zero
#ODOO-2 State Done

+++ app/Documents/Contract/OrderLine.php
@@ -32,7 +32,7 @@ class OrderLine {
 
     public float $nb_weeks;
     public int $nb_screens;
-    public int $quantity;
+    public float $quantity;
 
     public bool $is_production;
     public string $product;
